Add option to reassign reviewer
