+++ src/utils/constants.ts
@@ -33,6 +33,7 @@ export default {
   ASSIGNMENT_NOT_ASSIGNED: 'Not assigned',
   ASSIGNMENT_YOURSELF: 'Yourself',
   ASSIGNMENT_UNASSIGN: 'Un-assign',
+  ASSIGNMENT_REASSIGN: 'Re-assign',
   COLUMN_SERIAL: 'Serial',
   COLUMN_LAST_ACTIVE_DATE: 'Last Active',
   COLUMN_DEADLINE_DATE: 'Deadline',
@@ -124,6 +125,8 @@ export default {
   LABEL_REVIEW_REVIEWED_BY: 'Reviewed by',
   LABEL_REVIEW_CONSOLIDATION_BY: 'Consolidation by',
   LABEL_REVIEWER: 'Reviewer',
+  LABEL_REASSIGN_TO: 'Reassign to',
+  LABEL_UNASSIGN_FROM: 'Unassign from',
   LABEL_RESPONSE_UPDATE: 'Update',
   LABEL_RESPONSE_UPDATED: 'Updated',
   LABEL_SECTION_PROBLEM: 'Errors',
