+++ src/utils/data/assignmentOptions.ts
@@ -1,4 +1,5 @@
 export enum AssignmentOption {
+  REASSIGN = -2,
   UNASSIGN = -1,
   NOT_ASSIGNED = 0,
 }
